Bug fixes for input parameters

+++ code/Heimdall/src/main/java/unimore/t4/Heimdall/PreProcessing/LogProcessing.java
@@ -1,6 +1,7 @@
 package unimore.t4.Heimdall.PreProcessing;
 
 import java.io.File;
+import java.util.Objects;
 
 public class LogProcessing {
     /**
@@ -22,11 +23,17 @@ public class LogProcessing {
      * @param dirDstLogName  Stringa contenente le directory finale dei log files
      */
     public LogProcessing(String dirSrcLogName, String dirDstLogName, String dirDstJsonName){
-        preprocessingDirs(dirDstLogName, dirDstJsonName);
-        jsonWriter = new JsonWriter(dirDstJsonName);
-        logWriter = new LogWriter(dirDstLogName);
-        logParser = new LogParser(logWriter, jsonWriter);
-        logReader = new LogReader(dirSrcLogName, logParser);
+        if(Objects.equals(dirSrcLogName, "File_log")
+            && Objects.equals(dirDstLogName, "File_output")
+            && Objects.equals(dirDstJsonName, "File_Json")) {
+                preprocessingDirs(dirDstLogName, dirDstJsonName);
+                logParser = new LogParser(dirDstLogName, dirDstJsonName);
+                logReader = new LogReader(dirSrcLogName, logParser);
+        }
+        else {
+            System.err.println("Invalid directories names, " +
+                    "must be: File_log, File_output, File_Json");
+        }
     }
     /**
      *  processo che fa avviare il processo di lettura, parsing, scrittura
@@ -36,24 +43,34 @@ public class LogProcessing {
     }
 
     private void preprocessingDirs(String dirDstLogName, String dirDstJsonName) {
-        File dirDstJson = new File(new File("").getAbsolutePath() + File.separator
-                + dirDstJsonName);
-        if (!dirDstJson.mkdir()) {
-            boolean err = deleteDirectory(dirDstJson);
+        if(Objects.equals(dirDstLogName, "File_output") && Objects.equals(dirDstJsonName, "File_Json")){
+            File dirDstJson = new File(new File("").getAbsolutePath() + File.separator
+                    + dirDstJsonName);
+            if (!dirDstJson.mkdir()) {
+                boolean err = deleteDirectory(dirDstJson);
+            }
+            File dirDstLog = new File(new File("").getAbsolutePath() + File.separator
+                    + dirDstLogName);
+            if (!dirDstLog.mkdir()) {
+                boolean err = deleteDirectory(dirDstLog);
+            }
         }
-        File dirDstLog = new File(new File("").getAbsolutePath() + File.separator
-                + dirDstLogName);
-        if (!dirDstLog.mkdir()) {
-            boolean err = deleteDirectory(dirDstLog);
+        else {
+            System.err.println("Invalid directories names, " +
+                    "must be: File_output, File_Json");
         }
     }
     private boolean deleteDirectory(File directoryToBeDeleted) {
-        File[] allContents = directoryToBeDeleted.listFiles();
-        if (allContents != null) {
-            for (File file : allContents) {
-                deleteDirectory(file);
+        boolean isDirDeleted = false;
+        if (directoryToBeDeleted.exists()) {
+            File[] allContents = directoryToBeDeleted.listFiles();
+            if (allContents != null) {
+                for (File file : allContents) {
+                    deleteDirectory(file);
+                }
             }
+            isDirDeleted = directoryToBeDeleted.delete();
         }
-        return directoryToBeDeleted.delete();
+        return isDirDeleted;
     }
 }
